+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -28,6 +28,7 @@
 #include "EnumScopedCheck.h"
 #include "ForLoopCounterCheck.h"
 #include "FunctionNameUseCheck.h"
+#include "FunctionNoexceptCheck.h"
 #include "IdentifierTypographicallyUnambiguousCheck.h"
 #include "LambdaImplicitCaptureCheck.h"
 #include "LambdaParamListCheck.h"
@@ -104,8 +105,10 @@ public:
         "bsl-for-loop-counter");
     CheckFactories.registerCheck<FunctionNameUseCheck>(
         "bsl-function-name-use");
+    CheckFactories.registerCheck<FunctionNoexceptCheck>(
+        "bsl-function-noexcept");
     CheckFactories.registerCheck<IdentifierTypographicallyUnambiguousCheck>(
-        "bsl-identifier-typographically-unambiguous");
+        "bsl-identifier-typographically-unambiguous");    
     CheckFactories.registerCheck<LambdaImplicitCaptureCheck>(
         "bsl-lambda-implicit-capture");
     CheckFactories.registerCheck<LambdaParamListCheck>(